Testing effects
Getting effects set up (changes are from w/c 12 August 2024 - I didn't have stable internet access to commit them at the time of making these changes). These test out a delay effect (not very useful for long notes, not surprisingly), and various levels of reverb. The group opinion was that reverb is better, but a level between what I've here termed 'full reverb' (which is what I set it to for the rehearsal, which I rather like) and 'less reverb'. Actual level still to be determined.