Edits to touch controls
Red button now mutes and unmutes the sound without changing the volume. This works, but it requires a very quick tap on the touch bar to not cancel itself out, so while it seems to function perfectly well, it's a little unreliable as anything slightly too long won't mute/unmute, and anything too light risks not registering. I've rearranged the subpatch that controls the touch input with this. The yellow spot on the touch bar should now freeze the reverb if it gets out of hand, but I haven't yet tested it - trying to get the muting right, as I suspect freezing will require a similar approach.
Rounding abstraction hasn't been implemented in this version, but I keep wavering back and forth and found this one, so tracking that.