MUTE & VOL WORK!!
It's looking like I've solved the problems! Mute button & volume control are independent of one another; mute button triggers with the start of each press and it doesn't seem to matter how long you hold it for, it'll still toggle back and forth. It does miss some light taps, but I can investigate that further - that might just be a factor of needing to affix the bar a bit more solidly.