Changes to dry & wet signals
Deleted the dry signal from inside the phaser abstraction, adjusted the wet signal level to match that of the dry signal. Also changed the order of the signal addition and the reverb, so that the reverb is applied to the whole resultant signal, whether dry or with phaser, rather than the phased signal being added to the reverb signal. Basically, I think this is all working!