Add flash LED to reset
Added code to flash a second LED (turn on, then turn off again) when the reset zone is touched. I don't think this is strictly needed, but I'd like to add a second LED to ensure I understand the concepts behind doing this, not just following the instructions for one - this should make it easier to develop the project further in future.