UKRI submitted code
Cleaned-up version of the code for uploading to GitHub for Cyborg Soloists' annual UKRI report. I have deleted any lingering elements which are not being used and tested this. Note that if wanting to restore elements such as the reverb freeze, this is in earlier versions of the reverb_twig abstraction. This was the main change - otherwise mostly documenting the code with comments and removing a last print statement