Update README.md
Testing Claude.ai's integration with GitHub - this only seems to be able to upload files from GitHub, but I discovered that (contrary to my initial experiments) Claude IS able to interpret Pure Data files with the .pd extension, so long as you tell it that it's just a text file. What it can't do is change the files uploaded and recommit them, but it was able to summarise the patch I uploaded (the latest version of _main.pd) and provide it in Mardown format, ready to insert into README.md, as I have done here (with a couple of minor tweaks to wording and formatting, plus a statement clarifying Claude's role in creating this text). 
Verdict: The Claude GitHub integration is a good way of ensuring that I develop a workflow of regular updates to GitHub - if I'm attaching files I want it to look at from GitHub, then that means I'll be making commits more regularly to ensure I have the latest version on here to send to Claude, as opposed to my current system, which has been to do a Save As to save the file as a TXT then manually upload to Claude and sometimes forget to submit it. So the integration will support a more rigorous approach to version control as a form of documentation